requirements.txt
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
python-dotenv==1.0.0
Requests==2.31.0
